Bug 728272 - factor MockRecord out of TestServer11RepositorySession.
