+++ src/test/java/org/mozilla/android/sync/test/TestServer11RepositorySession.java
@@ -12,8 +12,8 @@ import java.net.URISyntaxException;
 import org.junit.Test;
 import org.mozilla.android.sync.test.helpers.BaseTestStorageRequestDelegate;
 import org.mozilla.android.sync.test.helpers.HTTPServerTestHelper;
+import org.mozilla.android.sync.test.helpers.MockRecord;
 import org.mozilla.android.sync.test.helpers.MockServer;
-import org.mozilla.gecko.sync.CryptoRecord;
 import org.mozilla.gecko.sync.Utils;
 import org.mozilla.gecko.sync.net.BaseResource;
 import org.mozilla.gecko.sync.net.SyncStorageRecordRequest;
@@ -54,32 +54,6 @@ public class TestServer11RepositorySession {
   static final String SYNC_KEY          = "eh7ppnb82iwr5kt3z3uyi5vr44";
 
   private HTTPServerTestHelper data     = new HTTPServerTestHelper();
-
-  public class MockRecord extends Record {
-    public MockRecord(String guid, String collection, long lastModified,
-        boolean deleted) {
-      super(guid, collection, lastModified, deleted);
-    }
-
-    @Override
-    public void initFromPayload(CryptoRecord payload) {
-    }
-
-    @Override
-    public CryptoRecord getPayload() {
-      return null;
-    }
-
-    @Override
-    public String toJSONString() {
-      return "{\"id\":\"" + guid + "\", \"payload\": \"foo\"}";
-    }
-
-    @Override
-    public Record copyWithIDs(String guid, long androidID) {
-      return new MockRecord(guid, this.collection, this.lastModified, this.deleted);
-    }
-  }
 
   public class MockServer11RepositorySession extends Server11RepositorySession {
     public MockServer11RepositorySession(Repository repository) {
